refactor: extract pure helpers from resumes.ts into resume_helpers.ts
Extract 10 pure utility functions (no Convex/Doc dependencies) into
resume_helpers.ts. Functions moved: toStringValue, toOptionalStringValue,
hasNonEmptyArray, readRecordArray, hasResumeFieldValue,
hasWorkHistoryDescriptionEntries, toRuleScores, resolveRuleScoreLookupKeys,
splitQueryTokens, matchesAllTokens.

resumes.ts re-exports all for backward compatibility. First pass of
decomposition (3,423L → 3,365L + 102L helpers).
